+++ NEWS.md
@@ -1,5 +1,6 @@
 # hutils 1.4.0
 * New functions:
+  - `%<->%` To swap values between objects
   - `average_bearing`, the bearing bisecting two vectors
   - `dir2`, (Windows only) a much faster version of `dir()`
   - `Mode`, statistical mode.
@@ -13,6 +14,7 @@
     used in `tidyr::uncount`.
   - `mutate_ntile` now works for a variable with `DT` 
   - `find_pattern_in` now accepts `file_contents_ignore_case`.
+  - `find_pattern_in` may 
 
 # hutils 1.3.0
 * Bug fixes: